feat: move ErrorBoundary to be around App base
